refactor: Abstract code block to separate module
Simplify automated testing. We mock `react-markdown` due to its ESM-only
architecture, so this module cannot be tested through the chat message
module.
